height & weight value get implements

+++ database/migrations/2022_03_21_131856_create_players_table.php
@@ -18,8 +18,8 @@ class CreatePlayersTable extends Migration
             $table->string('firstname')->comment('名前');
             $table->string('lastname')->comment('苗字');
             $table->string('birthday')->comment('生年月日');
-            $table->integer('height')->comment('身長');
-            $table->integer('weight')->comment('体重');
+            $table->double('height', 5, 2)->comment('身長');
+            $table->double('weight', 5, 2)->comment('体重');
             $table->string('image_file')->nullable()->comment('画像名');
             $table->timestamps();
         });
